% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_stats.R
\name{add_annual_low_flow}
\alias{add_annual_low_flow}
\title{add_annual_low_flow}
\usage{
add_annual_low_flow(flow.ts)
}
\arguments{
\item{flow.ts}{numeric USGS flow time series}
}
\value{
flow.ts.low numeric USGS annual low flow
}
\description{
Calculate annual low flow
}
