% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_stats.R
\name{add_annual_median_flow}
\alias{add_annual_median_flow}
\title{add_annual_median_flow}
\usage{
add_annual_median_flow(flow.ts)
}
\arguments{
\item{flow.ts}{numeric USGS flow time series}
}
\value{
flow.ts.med numeric USGS annual median flow
}
\description{
Calculate annual median flow
}
