% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_stats.R
\name{add_annual_peak_flow}
\alias{add_annual_peak_flow}
\title{add_annual_peak_flow}
\usage{
add_annual_peak_flow(flow.ts)
}
\arguments{
\item{flow.ts}{numeric USGS flow time series}
}
\value{
flow.ts.pk numeric USGS annual peak flow
}
\description{
Calculate annual peak flow
}
\examples{
\dontrun{library(dataRetrieval)
site_id <- '05114000'
parameterCd <- '00060'
startDate <- "2014-10-01"
endDate <- "2014-09-30"
flow.ts <- readNWISuv(site_id, parameterCd, startDate, endDate)
peak_flow <- flow.ts()
}

}
