% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{binconr}
\alias{binconr}
\title{Exact binomial confidence interval formatter}
\usage{
binconr(r, n, conf = 0.95, digits = 0, est = TRUE, method = "exact")
}
\arguments{
\item{r}{successes}

\item{n}{observations}

\item{conf}{confidence level}

\item{digits}{number of digits after decimal}

\item{est}{logical; if \code{TRUE}, the point estimate is also printed}

\item{method}{method to use; see \code{\link[Hmisc]{binconf}}}
}
\description{
Exact binomial confidence interval formatter
}
\examples{
binconr(5, 10, est = FALSE)
binconr(5, 10, conf = .9)
}

