% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pct_summaries.R
\name{calc_pct}
\alias{calc_pct}
\title{Percent of missing or bad data}
\usage{
calc_pct(ninterest, ntotal, digits = 2)
}
\arguments{
\item{ninterest}{vector amount of bad data}

\item{ntotal}{vector total amount of data}

\item{digits}{integer number of digits}
}
\examples{

data <- c(1,2,3,4,5)
data2 <- c(2,3,4,5,6)
percent <- calc_pct(data,data2)

}
