% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_water_year.R
\name{calc_water_year}
\alias{calc_water_year}
\title{calc_water_year}
\usage{
calc_water_year(date.vec)
}
\arguments{
\item{date.vec}{Vector with a string of dates}
}
\value{
Vector of the water year in which the given date occurs
}
\description{
Calculate the water year given a string of dates
}
\details{
Based on the USGS definition, a water year starts on October 1 of the year before, 
and ends on September 30. For example, water year 2015 started on 2014-10-01 and ended on 2015-09-30. 
See the USGS definition at https://water.usgs.gov/nwc/explain_data.html.
}
\examples{
mydates <- seq(as.POSIXct("2010-01-01"), as.POSIXct("2011-01-01"), by="day")
waterYear <- calc_water_year(date.vec = mydates)
}
