% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exercise.R
\name{check_current_exercise}
\alias{check_current_exercise}
\title{Check the current exercise}
\usage{
check_current_exercise(proj_path = getwd())
}
\arguments{
\item{proj_path}{path to the project folder, optional}
}
\description{
Check the current exercise
}
