% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exercise.R
\name{check_exercise}
\alias{check_exercise}
\title{Check an exercise file for completion}
\usage{
check_exercise(path)
}
\arguments{
\item{path}{path to the exercise file}
}
\value{
logical all tests ran successfully
}
\description{
Given the path to an exercise file, checks the file for completion by
ensuring all placeholder arguments are filled in and that all tests pass.
}
