% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exercise.R
\name{check_exercise_rstudio}
\alias{check_exercise_rstudio}
\title{Check a local exercise file for completed exercise}
\usage{
check_exercise_rstudio(proj_path = getwd())
}
\arguments{
\item{proj_path}{path to the project folder, optional}
}
\value{
a logical value, indicating whether the file passed all checks
}
\description{
This version of the file check is completed when the trainr project type is
"rstudio". Checks the "current_exercise.R" file for completion and outputs
messages to the console via the `cli` package.
}
