% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exercise.R
\name{check_exercise_shiny}
\alias{check_exercise_shiny}
\title{Check code lines for completed exercise in Shiny application}
\usage{
check_exercise_shiny(lines)
}
\arguments{
\item{lines}{character vector of exercise code lines}
}
\value{
a logical value, indicating whether the file passed all checks
}
\description{
This version of the file check is completed when the trainr project type is
"shiny". Checks the lines passed in as R code and outputs the results as
`message()`s, with `<span>`s for HTML formatting.
}
