% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exercise.R
\name{check_placeholders}
\alias{check_placeholders}
\title{Check an exercise for missing arguments}
\usage{
check_placeholders(file_lines)
}
\arguments{
\item{file_lines}{character vector of exercise file lines}
}
\value{
a list containing:
  - `success`: logical value indicating whether the check succeeded
  - `msg`: character value indicating the success/failure message
}
\description{
Exercise files contain a number of `?` markers that should be filled in by
the user. R interprets these as variables named `?`, which of course don't
exist. Given the lines (as character vectors) of the exercise file, raises
an informative error if any of those placeholders aren't filled in.
}
