% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exercise.R
\name{check_test_integrity}
\alias{check_test_integrity}
\title{Check to ensure exercise tests have not been altered}
\usage{
check_test_integrity(file_lines, test_lines)
}
\arguments{
\item{file_lines}{character vector of exercise file lines}

\item{test_lines}{character vector of original test lines}
}
\value{
a list containing:
  - `success`: logical value indicating whether the check succeeded
  - `msg`: character value indicating the success/failure message
}
\description{
Compares the tests from the user's modified exercise file to the tests
included in the source exercise file. Since the actual tests run are those
included in the source file, the results displayed for the user may not
match the results they would get if their tests were run.
}
