% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exercise.R
\name{check_tests}
\alias{check_tests}
\title{Check exercise file tests pass}
\usage{
check_tests(file_lines, proj_path = getwd())
}
\arguments{
\item{file_lines}{character vector of exercise file lines}

\item{proj_path}{path to the project folder, optional}
}
\value{
a list containing:
  - `success`: logical value indicating whether the check succeeded
  - `msg`: character value indicating the success/failure message
  - `output`: character vector of the lines of test output
}
\description{
Runs the R code, ensuring all tests pass
}
