% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-colour.R
\name{colour_time}
\alias{colour_time}
\title{Add colour to string
Add colour to string based on scheduled (\code{time}) and \code{expected} time of a
service.}
\usage{
colour_time(time, expected, show_colours = TRUE)
}
\arguments{
\item{time}{String with service scheduled time, format: \verb{\%H:\%M}.}

\item{expected}{String with service expected time, format: \verb{\%H:\%M}.}

\item{show_colours}{Boolean flag to indicate if the \code{Expected} time should
be colour coded or not, based on service time.}
}
\value{

}
\description{
Add colour to string
Add colour to string based on scheduled (\code{time}) and \code{expected} time of a
service.
}
\keyword{internal}
