% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exercise.R
\name{complete_current_exercise}
\alias{complete_current_exercise}
\title{Mark the current exercise as complete}
\usage{
complete_current_exercise(proj_path = getwd())
}
\arguments{
\item{proj_path}{path to the project folder, optional}
}
\description{
Set the current exercise metadata to indicate the exercise is complete and
save the current exercise to the 'completed' folder
}
