% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{config_get}
\alias{config_get}
\title{Get an option from the config file}
\usage{
config_get(option_name)
}
\arguments{
\item{option_name}{the name of the option to return}
}
\value{
the value of the named config option
}
\description{
Get an option from the config file
}
