% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{create.new.issues}
\alias{create.new.issues}
\title{Automatically create issues on the repo}
\usage{
create.new.issues(repo.name, issue.json = "inst/extdata/issuetemplates.json",
  org = "USGS-R", ctx = get.github.context())
}
\arguments{
\item{org}{string, GitHub organization to create repository. Defaults to "USGS-R"}

\item{ctx}{GitHub context for authentication, see \link[grithub]{get.github.context}}

\item{repo_name}{string, name for the new repository}

\item{issue_json}{file path indicating the JSON file to be used to define what 
issues to create. Defaults to the `issuetemplates.json` file in this package.}
}
\description{
This function should be run by instructors to setup the issues
that will be created for students in each class. The idea is that they fix and 
close out the issues in each instance of the course, and we reset the code to
have errors before the next course. We also need to reinstate the issues
associated with the errors. This function should automate that.
}
