% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exercise.R
\name{current_exercise_lines}
\alias{current_exercise_lines}
\title{Fetch the contents of the current exercise file}
\usage{
current_exercise_lines(proj_path = getwd())
}
\arguments{
\item{proj_path}{path to the project folder, optional}
}
\value{
(unmodified) lines from the current exercise
}
\description{
Fetch the contents of the current exercise file
}
