% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/trainr-package.R
\docType{data}
\name{demo}
\alias{demo}
\title{Patient demographics data}
\format{An object of class \code{data.frame} containing 50 observations and
16 variables:

\tabular{rll}{
\tab \code{id} \tab patient id number \cr
\tab \code{site} \tab enrollment site \cr
\tab \code{sex} \tab sex \cr
\tab \code{race} \tab race \cr
\tab \code{ethnic} \tab ethnicity \cr
\tab \code{d_dob} \tab day of birth (dd) \cr
\tab \code{m_dob} \tab month of birth (mm) \cr
\tab \code{y_dob} \tab year of birth (yyyy) \cr
\tab \code{d_reg} \tab day of registration (dd) \cr
\tab \code{m_reg} \tab month of registration (mm) \cr
\tab \code{y_reg} \tab year of registration (yyyy) \cr
\tab \code{subtype} \tab genetic subtype, activated bcell-like or germinal
center bcell-like \cr
\tab \code{ens} \tab number of extra nodal sites \cr
\tab \code{stage} \tab ann arbor disease stage \cr
\tab \code{ecog} \tab ecog performance status \cr
\tab \code{ldh} \tab ldh > ULN, normal or elevated \cr
}}
\usage{
demo
}
\description{
Baseline demographics data for patients enrolled on 15-000.
}
\seealso{
\code{\link{resp}}, \code{\link{surv}}, \code{\link{tox}}
\code{\link{blabs}}, \code{\link{elabs}}
}
\keyword{datasets}

