% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dmy}
\alias{dmy}
\title{Date parse}
\usage{
dmy(d, m, y, origin = c(1, 1, 1900))
}
\arguments{
\item{d, m, y}{day, month, year as single integers or vectors}

\item{origin}{vector of length 3 with origins for d, m, and y, respectively;
see details}
}
\value{
A vector of \code{\link{Date}}-formatted strings.
}
\description{
Parse day/month/year column data into standard date format.
}
\details{
For two-digit years, the origin year should be specified; otherwise, the
default of 1900 will be used. For NA year, month, or day, origin is used
for defaults, i.e., origin = c(15, 6, 2000) will convert missing days to
day 15, missing months to June, and missing years to 2000.
}
\examples{
dmy(25, 7, 87)
dmy(NA, NA, 2000:2005)

}

