% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainr-package.R
\docType{data}
\name{elabs}
\alias{elabs}
\title{End of treatment labs}
\format{An object of class \code{data.frame} containing 50 observations and
9 variables:

\tabular{rll}{
\tab \code{id} \tab patient id number \cr
\tab \code{timepoint} \tab timepoint \cr
\tab \code{ltd} \tab sum of largest tumor diameters (cm) \cr
\tab \code{wbc} \tab white blood cell count (k/mcL) \cr
\tab \code{hgb} \tab hemoglobin (g/dL) \cr
\tab \code{platelets} \tab platelets  (k/mcL) \cr
\tab \code{neutrophils} \tab neutrophils (\%) \cr
\tab \code{lymphocytes} \tab lymphocytes (\%%) \cr
\tab \code{monocytes} \tab monocytes (\%) \cr
}}
\usage{
elabs
}
\description{
End of treatment lab results for patients enrolled on 15-000.
}
\seealso{
\code{\link{demo}}, \code{\link{resp}}, \code{\link{surv}}
\code{\link{tox}}, \code{\link{blabs}}
}
\keyword{datasets}

