% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listing.R
\name{exercise_path_to_entry}
\alias{exercise_path_to_entry}
\title{Makes an entry for the exercise listing}
\usage{
exercise_path_to_entry(split_path, full_path)
}
\arguments{
\item{split_path}{a length-3 character vector containing the chapter name,
lesson name, and exercise name for an exercise, in that order}

\item{full_path}{the full path to an exercise file}
}
\value{
a one-row dataframe containing `chapter`, `lesson`, `exercise`,
`path`, and `completed`
}
\description{
Given a `split_path` and the `full_path` for an exercise, return a single
row dataframe containing an exercise listing entry for that exercise.
}
