% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract}
\alias{extract}
\alias{extract.GetServiceDetailsResult}
\alias{extract.GetStationBoardResult}
\alias{extract.busServices}
\alias{extract.callingPoint}
\alias{extract.callingPointList}
\alias{extract.trainServices}
\alias{extract.service}
\title{Extract data from XML request}
\usage{
extract(x, ...)

\method{extract}{GetServiceDetailsResult}(x, ...)

\method{extract}{GetStationBoardResult}(x, ...)

\method{extract}{busServices}(x, ...)

\method{extract}{callingPoint}(x, ...)

\method{extract}{callingPointList}(x, ..., class = "previousCallingPoints")

\method{extract}{trainServices}(x, ...)

\method{extract}{service}(x, ...)
}
\arguments{
\item{x}{List with request results.}

\item{...}{Optional parameters.}

\item{class}{String with class of the output object.}
}
\value{
Tibble with the appropriate format.
}
\description{
Extract data from XML request
}
\keyword{internal}
