% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainR.R
\name{get_calling_points}
\alias{get_calling_points}
\title{Obtain previous calling points}
\usage{
get_calling_points(data)
}
\arguments{
\item{data}{List with previous calling point records.}
}
\value{
Tibble with previous calling point records.
}
\description{
Obtain previous calling points
}
\keyword{internal}
