% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_crs}
\alias{get_crs}
\title{Get CRS code}
\usage{
get_crs(station, ...)
}
\arguments{
\item{station}{String with the station name, partial names are allowed.}

\item{...}{
  Arguments passed on to \code{\link[base:grep]{base::grep}}
  \describe{
    \item{\code{ignore.case}}{if \code{FALSE}, the pattern matching is \emph{case
      sensitive} and if \code{TRUE}, case is ignored during matching.}
    \item{\code{perl}}{logical.  Should Perl-compatible regexps be used?}
    \item{\code{value}}{if \code{FALSE}, a vector containing the (\code{integer})
    indices of the matches determined by \code{grep} is returned, and if
    \code{TRUE}, a vector containing the matching elements themselves is
    returned.}
    \item{\code{fixed}}{logical.  If \code{TRUE}, \code{pattern} is a string to be
    matched as is.  Overrides all conflicting arguments.}
    \item{\code{useBytes}}{logical.  If \code{TRUE} the matching is done
    byte-by-byte rather than character-by-character.  See
    \sQuote{Details}.}
    \item{\code{invert}}{logical.  If \code{TRUE} return indices or values for
    elements that do \emph{not} match.}
  }}
}
\value{
Vector of strings with the matched CRS codes, or \code{NULL} if no
match is found.
}
\description{
Get CRS code from a string with the station name.
}
\examples{
trainR::get_crs("London")
trainR::get_crs("Manchester")
trainR::get_crs("manchester", ignore.case = TRUE)
}
