% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listing.R
\name{get_current_exercise_listing}
\alias{get_current_exercise_listing}
\title{Get the listing for the current exercise}
\usage{
get_current_exercise_listing(proj_path = getwd())
}
\arguments{
\item{proj_path}{path to the project folder, optional}
}
\value{
a named list containing an exercise listing
}
\description{
The current exercise is considered to be the first exercise in the list where
'completed' == FALSE.
}
