% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_current_metadata}
\alias{get_current_metadata}
\title{Fetch the current exercise metadata}
\usage{
get_current_metadata(proj_path = getwd())
}
\arguments{
\item{proj_path}{path to the project folder, optional}
}
\value{
a list representing the current exercise metadata
}
\description{
Fetch the current exercise metadata
}
