% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_element}
\alias{get_element}
\title{Wrapper for \code{\link[base:getElement]{base::getElement}}}
\usage{
get_element(object, name, as_list = FALSE, default = NA)
}
\arguments{
\item{object}{
    object from which to extract element(s) or in which to replace element(s).
  }

\item{name}{
    A literal character string or a \link[base]{name} (possibly \link[base]{backtick}
    quoted).  For extraction, this is normally (see under
    \sQuote{Environments}) partially matched to the \code{\link[base]{names}}
    of the object.
  }

\item{as_list}{Boolean flag to return the element as a list.}

\item{default}{Default value to return if \code{name} is not found.}
}
\value{
Element from object, if exists, \code{default} otherwise.
}
\description{
Wrapper for \code{\link[base:getElement]{base::getElement}}, but instead of
returning \code{NULL} when the \code{name} is not found, returns an user
specified \code{default} value, \code{NA}.
}
\examples{
test_obj <- list(a = 4, c = 2)
get_element(test_obj, "a")
get_element(test_obj, "b")
}
