% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_location}
\alias{get_location}
\title{Get the full station name}
\usage{
get_location(crs)
}
\arguments{
\item{crs}{(string, 3 characters, alphabetic): The CRS code of the station.}
}
\value{
String with station name.
}
\description{
Get the full station name
}
\keyword{internal}
