% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exercise.R
\name{get_test_lines}
\alias{get_test_lines}
\title{Extract test lines from exercise lines}
\usage{
get_test_lines(lines)
}
\arguments{
\item{lines}{character vector of exercise code lines}
}
\value{
a character vector of lines containing test code
}
\description{
Extract test lines from exercise lines
}
