% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_token}
\alias{get_token}
\title{Get user's token}
\usage{
get_token(path = "inst/token.txt")
}
\arguments{
\item{path}{Path to text file with token.}
}
\value{
String with token, first line of the input file.
}
\description{
Get user's token to access the National Rail Enquiries (NRE) data feeds.
}
