% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainR.R
\name{GetArrDepBoardWithDetailsRequest}
\alias{GetArrDepBoardWithDetailsRequest}
\title{Get all public arrivals and departures}
\source{
Documentation for the Live Departure Boards Web Service
(LDBWS / OpenLDBWS):
\url{http://lite.realtime.nationalrail.co.uk/openldbws/}
}
\usage{
GetArrDepBoardWithDetailsRequest(
  crs,
  filterCrs = NA,
  filterType = "from",
  numRows = 150,
  timeOffset = 0,
  timeWindow = 120,
  token = get_token(),
  url = "https://lite.realtime.nationalrail.co.uk/OpenLDBWS/ldb11.asmx",
  verbose = FALSE
)
}
\arguments{
\item{crs}{(string, 3 characters, alphabetic): The CRS code of the location
for which the request is being made.}

\item{filterCrs}{(string, 3 characters, alphabetic): The CRS code of either
an origin or destination location to filter in. Optional.}

\item{filterType}{(string, either "from" or "to"): The type of filter to
apply. Filters services to include only those originating or terminating
at the \code{filterCrs} location. Defaults to "to".}

\item{numRows}{(integer, between 0 and 150 exclusive): The number of
services to return in the resulting station board.}

\item{timeOffset}{(integer, between -120 and 120 exclusive): An offset in
minutes against the current time to provide the station board for.
Defaults to 0.}

\item{timeWindow}{(integer, between -120 and 120 exclusive): How far into
the future in minutes, relative to \code{timeOffset}, to return services
for. Defaults to 120.}

\item{token}{Token to access the data feed. The token can be obtained at
\url{http://realtime.nationalrail.co.uk/OpenLDBWSRegistration/}.}

\item{url}{Data feed source URL.}

\item{verbose}{Boolean flag to indicate whether or not to show status
messages.}
}
\value{
Tibble with arrival and departure records. Each column is described
below:
\describe{
\item{generatedAt}{The time at which the station board was generated.}
\item{locationName}{The name of the location that the station board is for.}
\item{crs}{The CRS code of the location that the station board is for.}
\item{filterLocationName}{If a filter was requested, the location name of
the filter location.}
\item{filtercrs}{If a filter was requested, the CRS code of the filter
location.}
\item{filterType}{If a filter was requested, the type of filter.}
\item{nrccMessages}{An optional list of textual messages that should be
displayed with the station board. The message may include embedded and XML
encoded HTML-like hyperlinks and paragraphs. The messages are typically
used to display important disruption information that applies to the
location that the station board was for. Any embedded \code{<p>} tags are
used to force a new-line in the output. Embedded \code{<a>} tags allow
links to external web pages that may provide more information. Output
channels that do not support HTML should strip out the \code{<a>} tags and
just leave the enclosed text.}
\item{platformAvailable}{An optional value that indicates if platform
information is available. If this value is present with the value "true"
then platform information will be returned in the service lists. If this
value is not present, or has the value "false", then the platform "heading"
should be suppressed in the user interface for this station board.}
\item{areServicesAvailable}{An optional value that indicates if services
are currently available for this station board. If this value is present
with the value "false" then no services will be returned in the service
lists. This value may be set, for example, if access to a station has been
closed to the public at short notice, even though the scheduled services
are still running. It would be usual in such cases for one of the
\code{nrccMessages} to describe why the list of services has been
suppressed.}
\item{trainServices}{A list of \code{ServiceItem} object for each services
of the relevant type that is to appear on the station board. May contain
zero items, or may not be present at all.}
\item{busServices}{A list of \code{ServiceItem} object for each services
of the relevant type that is to appear on the station board. May contain
zero items, or may not be present at all.}
\item{ferryServices}{A list of \code{ServiceItem} object for each services
of the relevant type that is to appear on the station board. May contain
zero items, or may not be present at all.}
}
}
\description{
Get all public arrivals and departures for the supplied CRS code within a
defined time window, including service details.
}
\examples{
\dontrun{
rdg<- trainR::GetArrDepBoardWithDetailsRequest("RDG")
rdg <- trainR::GetArrDepBoardWithDetailsRequest("RDG", filterCrs = "BRI")
trainR::print(rdg)
}
}
\seealso{
Other OpenLDBSVWS requests: 
\code{\link{GetArrBoardRequest}()},
\code{\link{GetArrBoardWithDetailsRequest}()},
\code{\link{GetDepBoardRequest}()},
\code{\link{GetDepBoardWithDetailsRequest}()},
\code{\link{GetServiceDetailsRequest}()}
}
\concept{OpenLDBSVWS requests}
