% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{ipr}
\alias{ipr}
\title{Collapse using commas and "and"}
\usage{
ipr(x, sep = ",")
}
\arguments{
\item{x}{a vector}

\item{sep}{separator}
}
\description{
Collapse using commas and "and"
}
\examples{
ipr(1); ipr(1:2); ipr(1:5, sep = ';')
}

