% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exercise.R
\name{is_test}
\alias{is_test}
\title{Flag lines that contain test code}
\usage{
is_test(lines)
}
\arguments{
\item{lines}{character vector of exercise code lines}
}
\value{
a logical vector indicating which lines contain test code
}
\description{
This function assumes that tests begin with the line containing a 'flag' and
continue until the end of the file.
}
