% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{is_trainr_project}
\alias{is_trainr_project}
\title{Is this path a *trainr* project?}
\usage{
is_trainr_project(proj_path = getwd())
}
\arguments{
\item{proj_path}{path to the potential project folder}
}
\value{
logical path is a *trainr* project path
}
\description{
If the directory at the path contains both the current exercise metadata and
the exercise listings, then it is a *trainr* project folder.
}
