% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_valid_crs}
\alias{is_valid_crs}
\title{Validate station code (CRS)}
\usage{
is_valid_crs(x, parameter = "crs")
}
\arguments{
\item{x}{Station code.}

\item{parameter}{String with the name of the parameter to check.}
}
\value{
Nothing, call for its side effect.
}
\description{
Validate station code (CRS)
}
\keyword{internal}
