% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exercise.R
\name{mark_current_exercise_complete}
\alias{mark_current_exercise_complete}
\title{Update the completed status of the current exercise}
\usage{
mark_current_exercise_complete(proj_path = getwd())
}
\arguments{
\item{proj_path}{path to the project folder, optional}
}
\description{
Update the completed status of the current exercise
}
