% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exercise.R
\name{missing_arguments}
\alias{missing_arguments}
\title{Alert the user if their exercise is missing any arguments}
\usage{
missing_arguments(file_lines)
}
\arguments{
\item{file_lines}{character vector of exercise file lines}
}
\value{
logical there are missing arguments
}
\description{
Exercise files contain a number of `?` markers that should be filled in by
the user. R interprets these as variables named `?`, which of course don't
exist. Given the lines (as character vectors) of the exercise file, raises
an informative error if any of those placeholders aren't filled in.
}
