% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msg.R
\name{msg_alert_danger}
\alias{msg_alert_danger}
\title{Message Alerts}
\usage{
msg_alert_danger(msg)
}
\arguments{
\item{msg}{text of the message}
}
\description{
Displays a short status message with HTML formatting
}
