% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msg.R
\name{msg_alert_warning}
\alias{msg_alert_warning}
\title{Message Alerts}
\usage{
msg_alert_warning(msg)
}
\arguments{
\item{msg}{text of the message}
}
\value{
an HTML formatted message
}
\description{
Displays a short status message with HTML formatting
}
