% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msg.R
\name{msg_h1}
\alias{msg_h1}
\title{Message Headings}
\usage{
msg_h1(msg)
}
\arguments{
\item{msg}{text of the heading}

\item{out.width}{width of the output line, in characters}
}
\value{
an HTML formatted message
}
\description{
Outputs a message with the specified heading format
}
