% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{new_exercise_listing}
\alias{new_exercise_listing}
\title{Generate a fresh exercise listing}
\usage{
new_exercise_listing()
}
\value{
a dataframe including an entry for each exercise
}
\description{
Each listing entry contains the `chapter` name, `lesson` name,
`exercise` name, `path` to the exercise file, and a boolean indicating if
that exercise has been `completed`. Generated from the contents of the
'inst/lessons' folder.
}
