% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exercise.R
\name{original_test_lines}
\alias{original_test_lines}
\title{Get test code from original exercise file}
\usage{
original_test_lines(proj_path = getwd())
}
\arguments{
\item{proj_path}{path to the project folder, optional}
}
\value{
a character vector containing the test code for the current exercise
}
\description{
Fetches the lines containing test code from the original, unaltered version
of the current exercise file. Current exercise file is the one listed in the
exercise listing in `proj_path`.
}
