% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pct_summaries.R
\name{pct_censored}
\alias{pct_censored}
\title{Censored Data}
\usage{
pct_censored(vals, digits = 2)
}
\arguments{
\item{vals}{vector amount of bad censored}
}
\description{
Flags data censored.
}
\examples{
data <- c(1,2,3,4,5,16,34)
censored <- pct_censored(data)



}
