% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pct_summaries.R
\name{pct_missing}
\alias{pct_missing}
\title{Missing Data}
\usage{
pct_missing(vals, digits = 2)
}
\arguments{
\item{vals}{vector amount of missing data}
}
\description{
Flags missing data
}
\examples{

data <- c(1,2,3,4,5,NA)
missing <- pct_missing(data)
}
