% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print}
\alias{print}
\alias{print.GetStationBoardResult}
\alias{print.busServices}
\alias{print.trainServices}
\alias{print.previousCallingPoints}
\alias{print.subsequentCallingPoints}
\title{Print Values}
\usage{
print(x, ...)

\method{print}{GetStationBoardResult}(x, ...)

\method{print}{busServices}(x, ...)

\method{print}{trainServices}(x, ...)

\method{print}{previousCallingPoints}(x, ...)

\method{print}{subsequentCallingPoints}(x, ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Input data invisibly.
}
\description{
Print Values
}
