% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print_board}
\alias{print_board}
\title{Print arrivals/departures board}
\usage{
print_board(
  x,
  show_details = FALSE,
  station = "destination",
  string = FALSE,
  show_header = TRUE,
  ...
)
}
\arguments{
\item{x}{Tibble with arrivals/departures information.}

\item{show_details}{Boolean flag to indicate if detail information about
previous calling points should be included or not.}

\item{station}{String to indicate if the destination or origin station
should be displayed.}

\item{string}{Boolean flag to indicate whether or not the station board
results, should be returned as a string.}

\item{show_header}{Boolean flag to indicate if the header of board should be
displayed or not.}

\item{...}{Optional parameters (not used).}
}
\value{
Nothing, call for its side effect.
}
\description{
Print arrivals/departures board
}
\keyword{internal}
