% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print_counts}
\alias{print_counts}
\title{Count formatter}
\usage{
print_counts(x, ...)
}
\arguments{
\item{x}{a vector of character strings (or factors)}

\item{sep}{separator; see \code{\link{ipr}}}
}
\description{
Formats and prints vectors with Ns and percents.
}
\examples{
table(tx <- rep(c('RCHOP','R-CVP','RCHOEP'), c(10, 6, 4)))
print_counts(tx)

}

