% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{project_file_path}
\alias{project_file_path}
\title{Get the path for a local project file}
\usage{
project_file_path(proj_path, file_type)
}
\arguments{
\item{proj_path}{path to the current project folder}

\item{file_type}{the type of project file, as named in the config}
}
\value{
a file path (character)
}
\description{
Returns the path to a given type of *trainr* project file. Does not check
to see if that file exists, does check to be sure the project path is valid.
}
