% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reclass}
\alias{reclass}
\title{Update class of data object}
\usage{
reclass(data, class)
}
\arguments{
\item{data}{Data object (e.g. \code{list}).}

\item{class}{String of new class.}
}
\value{
Original data object with new \code{class}.
}
\description{
Update class of data object
}
\examples{
out <- list(woof = list(name = "Barto", age = 6)) \%>\%
  trainR:::reclass(names(.))
}
\keyword{internal}
