% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msg.R
\name{replace_console_color_codes}
\alias{replace_console_color_codes}
\title{Replace console color codes in text with styled spans}
\usage{
replace_console_color_codes(lines)
}
\arguments{
\item{lines}{lines of text to make replacements in}
}
\value{
lines of text with replacements made
}
\description{
The current replacements are:
- \033[1m  -> '<span style = "font-weight: bold;">'
- \033[32m -> '<span style = "color: limegreen">'
- \033[33m -> '<span style = "color: crimson">'
- \033[90m -> '<span style = "color: darkslategray">'
- \033[39m -> '</span>'
- \033[22m -> '</span>'
}
