% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainR.R
\name{request}
\alias{request}
\title{Submit a SOAP XML request}
\usage{
request(body, header, url, verbose = FALSE, class = NULL)
}
\arguments{
\item{body}{XML body arguments.}

\item{header}{XML header arguments.}

\item{url}{Data feed source URL.}

\item{verbose}{Boolean flag to indicate whether or not to show status
messages.}

\item{class}{String with request class.}
}
\value{
Request results or error message if not found.
}
\description{
Submit a SOAP XML request
}
\keyword{internal}
