% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/trainr-package.R
\docType{data}
\name{resp}
\alias{resp}
\title{Response data}
\format{An object of class \code{data.frame} containing 50 observations and
5 variables:

\tabular{rll}{
\tab \code{id} \tab patient id number \cr
\tab \code{resp} \tab response \cr
\tab \code{d_resp} \tab day of response assessment (dd) \cr
\tab \code{m_resp} \tab month of response assessment (mm) \cr
\tab \code{y_resp} \tab year of response assessment (yyyy) \cr
}}
\usage{
resp
}
\description{
Response assessments for patients enrolled on 15-000.
}
\seealso{
\code{\link{demo}}, \code{\link{tox}}, \code{\link{surv}},
\code{\link{blabs}}, \code{\link{elabs}}
}
\keyword{datasets}

