% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmse.R
\name{rmse.by.year}
\alias{rmse.by.year}
\title{Root mean squared error (rmse) by year}
\usage{
rmse.by.year(df)
}
\arguments{
\item{df}{data frame with columns named "year", "obs", and "pred".}
}
\value{
returns a rmse value for each year
}
\description{
calculates the rmse between predicted and observed values for each year in a dataset.
}
\examples{
year <- c(rep(1,25),rep(2,25),rep(3,25))
obs <- c(1:75)
e <- c(rnorm(n=25,mean=0,sd=2),rnorm(n=25,mean=0,sd=10),rnorm(n=25,mean=0,sd=50))
pred <- obs + e
df <- data.frame(year,obs,pred)
rmse.by.year(df)
}
