% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmse.R
\name{rmse}
\alias{rmse}
\title{Root mean squared error (rmse)}
\usage{
rmse(obs, pred)
}
\arguments{
\item{obs}{observed values}

\item{pred}{predicted values}
}
\value{
returns a single rmse value
}
\description{
calculates the rmse between predicted and observed values.
}
\examples{
obs <- 1:50
e <- rnorm(n=50,mean=0,sd=5)
pred <- obs + e
rmse(obs,pred)
plot(obs,pred)
}
