% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exercise.R
\name{save_completed_exercise_file}
\alias{save_completed_exercise_file}
\title{Save current exercise to 'completed' folder}
\usage{
save_completed_exercise_file(proj_path = getwd())
}
\arguments{
\item{proj_path}{path to the project folder, optional}
}
\description{
Maintains a folder of completed exercises, copies the currently completed
exercise file into the completed folder.
}
