% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup_rprofile}
\alias{setup_rprofile}
\title{Setup .Rprofile file}
\usage{
setup_rprofile(proj_path)
}
\arguments{
\item{proj_path}{}
}
\description{
We want to load `trainr` and update the exercise listing on project open
}
