% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup_rstudio_dir}
\alias{setup_rstudio_dir}
\title{Function stub for creating a new RStudio project}
\usage{
setup_rstudio_dir(proj_path)
}
\arguments{
\item{proj_path}{path to the new project directory}
}
\description{
Function stub for creating a new RStudio project
}
