% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{station_codes}
\alias{station_codes}
\title{National Rail Enquiries (NRE) Station Codes dataset}
\format{
A data frame with 2580 rows and 2 variables:
\describe{
\item{name}{Station name.}
\item{crs}{Station code.}
}
}
\source{
\url{https://www.nationalrail.co.uk/stations_destinations/48541.aspx}
}
\usage{
data(station_codes)
}
\description{
A dataset containing information of 2580 British National Rail Enquiries
(NRE) station 3 letter alpha (CRS) codes.
}
\author{
National Rail Enquiries (NRE)
}
\keyword{datasets}
