% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainr-package.R
\docType{data}
\name{surv}
\alias{surv}
\title{Survival data}
\format{An object of class \code{data.frame} containing 50 observations and
5 variables:

\tabular{rll}{
\tab \code{id} \tab patient id number \cr
\tab \code{status} \tab status, alive or dead \cr
\tab \code{d_status} \tab day of survival assessment (dd) \cr
\tab \code{m_status} \tab month of survival assessment (mm) \cr
\tab \code{y_status} \tab year of survival assessment (yyyy) \cr
}}
\usage{
surv
}
\description{
Survival assessments for patients enrolled on 15-000.
}
\seealso{
\code{\link{demo}}, \code{\link{tox}}, \code{\link{resp}}
\code{\link{blabs}}, \code{\link{elabs}}
}
\keyword{datasets}

