% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{surv_table}
\alias{surv_table}
\title{survfit summary table}
\usage{
surv_table(s, digits = 3, times = pretty(range(s$time)), ...)
}
\arguments{
\item{s}{an object of class \code{\link{survfit}}}

\item{digits}{number of digits after decimal}

\item{times}{vector of survival times}

\item{\dots}{additional arguments passed to \code{\link{summary.surfit}}}
}
\description{
Prints a formatted summary table with estimates and confidence intervals
for \code{\link{suvfit}} objects.
}
\examples{
\dontrun{
library('survival')
data(cancer)
fit <- survfit(Surv(time, status) ~ 1, data = cancer, conf.type = 'log-log')
surv_table(fit, times = 0:4 * 50, digits = 2)
}
}

