% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_exerecise.R
\name{to_exercise}
\alias{to_exercise}
\title{Convert a reproducible 'rmarkdown' document to 'learnr' exercises.}
\usage{
to_exercise(from, to)
}
\arguments{
\item{from}{Path to read lines from, maybe a URL.}

\item{to}{Path to write the output into.}
}
\value{
Invisible \code{from}.
}
\description{
Convert a reproducible 'rmarkdown' document to 'learnr' exercises.
}
\examples{
url <- "https://raw.githubusercontent.com/maurolepore/trainr/dump-features/README.Rmd"
# Look before
writeLines(readLines(url))

path <- tempfile(fileext = ".Rmd")
to_exercise(url, path)
# Look after
writeLines(readLines(path))

if (interactive()) rmarkdown::run(path)
}
