% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainr-package.R
\docType{data}
\name{tox}
\alias{tox}
\title{Toxicity data}
\format{An object of class \code{data.frame} containing 50 observations and
3 variables:

\tabular{rll}{
\tab \code{id} \tab patient id number \cr
\tab \code{tox_code} \tab toxicity code (CTCAE v4) \cr
\tab \code{tox_grade} \tab toxicity grade \cr
}}
\usage{
tox
}
\description{
Treatment-related toxicity data for patients enrolled on 15-000.
}
\seealso{
\code{\link{demo}}, \code{\link{resp}}, \code{\link{surv}}
\code{\link{blabs}}, \code{\link{elabs}}
}
\keyword{datasets}

