% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{tox_worst}
\alias{tox_worst}
\title{Worst toxicities}
\usage{
tox_worst(dat, id = "id", tox_desc = "tox_desc", tox_grade = "tox_grade")
}
\arguments{
\item{dat}{toxicity data}

\item{id}{id variable}

\item{tox_desc}{toxicity description}

\item{tox_grade}{toxicity grade as a factor}
}
\value{
A list of length three with worst toxicities, the original data frame, and
indices of duplicates removed from \code{dat}.
}
\description{
Get worst toxicity grades by subject, toxicity.
}

