% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch.R
\name{tr_categorical_analysis}
\alias{tr_categorical_analysis}
\title{Categorical data analysis}
\usage{
tr_categorical_analysis()
}
\description{
Launch categorical data analysis tutorial.
}
\examples{
\dontrun{
tr_categorical_analysis()
}

}
