% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch.R
\name{tr_date_manipulation}
\alias{tr_date_manipulation}
\title{Date manipulation}
\usage{
tr_date_manipulation()
}
\description{
Launch date and time manipulation tutorial.
}
\examples{
\dontrun{
tr_date_manipulation()
}

}
