% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{tr_download_data}
\alias{tr_download_data}
\title{Download datasets}
\usage{
tr_download_data()
}
\description{
Download data used in the tutorials from GitHub repository.
}
\examples{
tr_download_data()

}
