% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainR.R
\docType{package}
\name{trainR-package}
\alias{trainR}
\alias{trainR-package}
\title{trainR: An Interface to the National Rail Enquiries Systems}
\description{
The goal of trainR is to provide a simple interface to the National 
    Rail Enquiries (NRE) systems.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/villegar/trainR/}
  \item \url{https://villegar.github.io/trainR/}
  \item Report bugs at \url{https://github.com/villegar/trainR/issues/}
}

}
\author{
\strong{Maintainer}: Roberto Villegas-Diaz \email{villegas.roberto@hotmail.com} (\href{https://orcid.org/0000-0001-5036-8661}{ORCID})

}
\keyword{internal}
