% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exercise.R
\name{update_current_exercise}
\alias{update_current_exercise}
\title{Update the current exercise}
\usage{
update_current_exercise(proj_path = getwd())
}
\arguments{
\item{proj_path}{}
}
\value{
a dataframe containing the exercise listing
}
\description{
If the current exercise has been completed, mark the next exercise in sequence
as the current exercise.
}
