% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exercise.R
\name{update_exercise_entry}
\alias{update_exercise_entry}
\title{Update the completed status of the current exercise}
\usage{
update_exercise_entry(proj_path = getwd())
}
\arguments{
\item{proj_path}{path to the project folder, optional}
}
\value{
proj_path path to the project folder, optional
}
\description{
Update the completed status of the current exercise
}
