% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{update_exercise_listing}
\alias{update_exercise_listing}
\title{Create a new exercise listing if none exists, otherwise update the
existing listing with the package file structure}
\usage{
update_exercise_listing(proj_path)
}
\arguments{
\item{proj_path}{path to create the exercise listing in}
}
\value{
path to the exercise listing
}
\description{
Create a new exercise listing if none exists, otherwise update the
existing listing with the package file structure
}
