% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate}
\alias{validate}
\title{Validate output from \code{request}}
\usage{
validate(data)
}
\arguments{
\item{data}{Data object (e.g. \code{list}).}
}
\value{
Original data object, if data is valid
}
\description{
Validate output from \code{request}
}
\keyword{internal}
