% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{write_config_file}
\alias{write_config_file}
\title{Write `trainr` config to config file}
\usage{
write_config_file(proj_path, config)
}
\arguments{
\item{proj_path}{path to the new project directory}

\item{config}{list of config options}
}
\description{
Write `trainr` config to config file
}
