% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{write_current_exercise_file}
\alias{write_current_exercise_file}
\title{Copy the first exercise to current exercise file}
\usage{
write_current_exercise_file(proj_path)
}
\arguments{
\item{proj_path}{path to the new project directory}
}
\description{
Copy the first exercise to current exercise file
}
