% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{write_exercise_listing}
\alias{write_exercise_listing}
\title{Write a new exercise listing file}
\usage{
write_exercise_listing(proj_path)
}
\arguments{
\item{proj_path}{path to the new project directory}
}
\description{
Write a new exercise listing file
}
