% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{write_metadata}
\alias{write_metadata}
\title{Write current exercise metadata to file}
\usage{
write_metadata(proj_path = getwd(), metadata)
}
\arguments{
\item{proj_path}{path to the project folder, optional}

\item{metadata}{current exercise metadata object}
}
\value{

}
\description{
Write current exercise metadata to file
}
