% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_tutorials.R
\name{write_tutorial}
\alias{write_tutorial}
\title{Convert a reproducible 'rmarkdown' document to 'learnr' exercises.}
\usage{
write_tutorial(from, to)
}
\arguments{
\item{from}{Path to read lines from, maybe a URL.}

\item{to}{Path to write the output into.}
}
\value{
Invisible \code{from}.
}
\description{
Convert a reproducible 'rmarkdown' document to 'learnr' exercises.
}
\examples{
url <- "https://raw.githubusercontent.com/r-lib/fs/master/README.Rmd"
path <- tempfile(".Rmd")
write_tutorial(url, path)
# Look
writeLines(head(readLines(path), 10))
}
