% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{binconr}
\alias{binconr}
\title{Exact binomial confidence interval formatter}
\usage{
binconr(r, n, conf = 0.95, digits = 0, est = TRUE, method = "exact")
}
\arguments{
\item{r}{successes}

\item{n}{observations}

\item{conf}{confidence level}

\item{digits}{number of digits after decimal}

\item{est}{logical; if \code{TRUE}, the point estimate is also printed}

\item{method}{method to use; see \code{\link[Hmisc]{binconf}}}
}
\description{
Calculates exact binomial confidence interval (\code{\link[Hmisc]{binconf}})
and prints formatted string.
}
\examples{
binconr(5, 10, est = FALSE)
binconr(5, 10, conf = .9)

}
\seealso{
\code{\link[Hmisc]{binconf}}; \code{desmon::binci}
}

