% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exercise.R
\name{check_exercise_shiny}
\alias{check_exercise_shiny}
\title{Check code lines for completed exercise in Shiny application}
\usage{
check_exercise_shiny(lines, proj_path = getwd())
}
\arguments{
\item{lines}{character vector of exercise code lines}

\item{proj_path}{path to the project folder, optional}
}
\value{
a list containing:
  `success` a boolean indicating whether the code was evaluated successfully
            and passed all tests
  `msg`     a character vector of HTML formatted messages generated while
            evaluating the code
}
\description{
This version of the file check is completed when the trainr project type is
"shiny". Checks the lines passed in as R code and outputs the results as
`message()`s, with `<span>`s for HTML formatting.
}
