% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainR.R
\name{extract}
\alias{extract}
\title{Extract departure details}
\usage{
extract(x)
}
\arguments{
\item{x}{List with departure details.}
}
\value{
Tibble with departure details.
}
\description{
Extract departure details
}
\keyword{internal}
