% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_token}
\alias{get_token}
\title{Get user's token}
\usage{
get_token(ENV = "NRE")
}
\arguments{
\item{ENV}{String with environment variable containing the token to access
the NRE data feeds (default = "NRE").}
}
\value{
String with token.
}
\description{
Get user's token to access the National Rail Enquiries (NRE) data feeds.
}
