% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ipr}
\alias{ipr}
\title{Collapse vectors for printing}
\usage{
ipr(x, sep = ",")
}
\arguments{
\item{x}{a vector}

\item{sep}{separator}
}
\description{
Collapse a vector using commas and "and" for readability.
}
\examples{
ipr(1)
ipr(1:2)
ipr(1:5, sep = ';')

ipr(Vectorize(num2char)(1:5, cap = FALSE))

}

