% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exercise.R
\name{next_exercise}
\alias{next_exercise}
\title{Advance to the next exercise file}
\usage{
next_exercise(proj_path = getwd())
}
\arguments{
\item{proj_path}{path to the project folder, optional}
}
\value{
an object representing the current exercise
}
\description{
Updates (creates) the current exercise file to be the current exercise
in the exercise listing
}
