% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{num2char}
\alias{num2char}
\title{Numeric to character string}
\usage{
num2char(num, informal = FALSE, cap = TRUE)
}
\arguments{
\item{num}{integer in [-999, 999]}

\item{informal}{logical; if \code{TRUE}, adds "and" before tens or ones}

\item{cap}{logical; if \code{TRUE}, capitalized first word}
}
\description{
Numeric to character string
}
\examples{
num2char(28, informal = TRUE)

v_num2char <- Vectorize(num2char)
nums <- c(-100, 110, 322, 012, 201, -152, 4)
v_num2char(nums)
}

