% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ts.R
\name{plot_ts}
\alias{plot_ts}
\title{Timeseries plotting function}
\usage{
plot_ts(xdat, ydat, y.axis.label = "Y Value")
}
\arguments{
\item{xdat}{data for the x axis}

\item{ydat}{data for the y axis}

\item{y.axis.label}{name to put on the y axis, defaults to "Y Value"}
}
\value{
plot of timeseries
}
\description{
Plot a timeseries by giving the x and y data.
}
