% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print}
\alias{print}
\alias{print.GetStationBoardResult}
\alias{print.busServices}
\alias{print.trainServices}
\alias{print.previousCallingPoints}
\title{Print data}
\usage{
print(x, ...)

\method{print}{GetStationBoardResult}(x, ...)

\method{print}{busServices}(x, ...)

\method{print}{trainServices}(x, ...)

\method{print}{previousCallingPoints}(x, ...)
}
\arguments{
\item{x}{Data}

\item{...}{Optional parameters.}
}
\value{
Input data invisibly.
}
\description{
Print data
}
