% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print}
\alias{print}
\alias{print.StationBoard}
\alias{print.busServices}
\alias{print.trainServices}
\alias{print.previousCallingPoints}
\alias{print.subsequentCallingPoints}
\title{Print Values}
\usage{
print(x, ...)

\method{print}{StationBoard}(x, ..., station = NA, string = FALSE)

\method{print}{busServices}(x, ...)

\method{print}{trainServices}(x, ...)

\method{print}{previousCallingPoints}(x, ...)

\method{print}{subsequentCallingPoints}(x, ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{...}{further arguments passed to or from other methods.}

\item{station}{String to indicate if the destination or origin station
should be displayed.}

\item{string}{Boolean flag to indicate whether or not the station board
results, should be returned as a string.}
}
\value{
Input data invisibly.
}
\description{
Print Values
}
