% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print_board}
\alias{print_board}
\title{Print arrivals/departures board}
\usage{
print_board(x, show_details = FALSE, station = "destination", ...)
}
\arguments{
\item{x}{Tibble with arrivals/departures information.}

\item{show_details}{Boolean flag to indicate if detail information about
previous calling points should be included or not.}

\item{station}{String to indicate if the destination or origin station
should be displayed.}

\item{...}{Optional parameters (not used).}
}
\value{
Nothing, call for its side effect.
}
\description{
Print arrivals/departures board
}
\keyword{internal}
