% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_token}
\alias{set_token}
\title{Configure user's token}
\usage{
set_token(path = "inst/token.txt")
}
\arguments{
\item{path}{Path to text file where the token should be stored.}
}
\description{
Configure user's token to access the National Rail Enquiries (NRE) data
feeds.
}
