% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msg.R
\name{text_to_html}
\alias{text_to_html}
\title{Replace console color codes in text with styled spans}
\usage{
text_to_html(lines)
}
\arguments{
\item{lines}{lines of text to make replacements in}
}
\value{
lines of text with replacements made
}
\description{
The current replacements are:
- \033[1m  -> '<span style = "font-weight: bold;">'
- \033[32m -> '<span style = "color: green">'
- \033[33m -> '<span style = "color: red">'
- \033[90m -> '<span style = "color: magenta">'
- \033[39m -> '</span>'
- \033[22m -> '</span>'
}
